More filtering stuff.

+++ common/raw_archive_db.py
@@ -4,6 +4,7 @@ from sqlalchemy import Table
 
 from sqlalchemy import Column
 from sqlalchemy import BigInteger
+from sqlalchemy import Integer
 from sqlalchemy import Text
 from sqlalchemy import Float
 from sqlalchemy import Boolean
@@ -40,13 +41,13 @@ class RawWebPages(common.db_base.Base):
 
 	id                = Column(BigInteger, primary_key = True, index = True)
 	state             = Column(common.db_types.dlstate_enum, default='new', index=True, nullable=False)
-	errno             = Column(BigInteger, default='0')
+	errno             = Column(Integer, default='0')
 	url               = Column(Text, nullable = False, index = True, unique = True)
 	starturl          = Column(Text, nullable = False)
 	netloc            = Column(Text, nullable = False, index = True)
 
-	priority          = Column(BigInteger, default=1000000, index=True, nullable=False)
-	distance          = Column(BigInteger, index=True, nullable=False)
+	priority          = Column(Integer, default=1000000, index=True, nullable=False)
+	distance          = Column(Integer, index=True, nullable=False)
 
 	mimetype          = Column(Text)
